images: Vector to use package, use in Consul
We were previously writing out the Vector configuration at start time
which made it hard to debug the Vector work since things were
overwritten at restart. Since Vector can interpolate variables, instead
use the systemd file that Vector comes with and overwrite the
/etc/default section with the appropriate variables.

This also adds an exponentional backoff to the finding the logs
aggregator instead of just ignoring it. We shouldn't come up without
getting a connection to the logs aggregator since if we can't see the
logs aggregator than this is as good as failed.

As well, this moves our Consul servers to use the newly refactored
Vector configuration so we can get logs from Consul servers.

Change-Id: Ia74d9a098e3b15190d82e1b7bff3b80b71d08a61
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2072
Tested-by: Buildkite CI
Reviewed-by: Marcus Bastian <marcus@readyset.io>