Drop cached queries when overwritting with same name
When overwritting a cached query that was previously migrated with the
same name as a new cached query, drop the previous cached query first.

This resolves a bug where running:
1. 'CREATE CACHED QUERY q AS SELECT a FROM t'
2. 'CREATE CACHED QUERY q AS SELECT b FROM t'
3. 'SELECT a FROM t'
would return results for 'SELECT b FROM t'.

This behavior assumes that the migration for the overwritten query is
no longer desired.

Change-Id: If846441c631116c060d97637566acfb609e3cee8
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2086
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>