ops/docker: remove fragile pin for libssl-dev apt package
We keep seeing broken builds due to how we're pinning the libssl-dev apt
package. It would appear that every couple of weeks a new version is
released which effectively breaks all our container image builds.

To avoid that, this CL aims to remove the pin and take the latest
version. This is basically what we were trying to do with the pin
initially, but weren't anticipating package versions to be removed so
frequently.

Error message that this CL aims to resolve:

> 8 26.13 E: Version '1.1.1f-1ubuntu2.10' for 'libssl-dev' was not found

Change-Id: I3017c9f49e305c96d0a1e5abe23246f45391b7a9
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2091
Reviewed-by: Griffin Smith <griffin@readyset.io>
Tested-by: Griffin Smith <griffin@readyset.io>