docs: Start a page on replay paths + key provenance
To prepare for the (unreasonable amount of) complexity I'm about to add
to replay paths to re-add support for partial straddled joins, add the
beginning of a page on replay paths + key provenance, explaining how
both work and why they're necessary up to the point of what existed in
the PhD thesis.

Next up, I'll be adding in a section about *remapped* column
provenance - I've left in a section here with a TODO about that.

Change-Id: I27ee92c013d60d94531d9abd25895cd646ea132c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2095
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>