Relax hash-equal condition for hash laws
The rules of the rust Hash trait are that if two values are equal, their
hashes must also be equal. However, our macro for generating tests for
the laws of this trait were more strict - they also required that if two
values are *inequal*, their hashes must also be inequal, but this is
impossible to achieve due to collisions. This eventually got caught in
CI (thanks proptest!). This change relaxes the test case to make it more
closely resemble the actual rules of the trait.

Change-Id: I56d2e7bd966b7ed29709194b8b9d9e5317d8319d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2104
Reviewed-by: Harley Klein <harley@readyset.io>
Reviewed-by: Justin Miron <justin@readyset.io>
Tested-by: Buildkite CI