dataflow: Bring back partial straddled joins
Bring back upquery remapping for straddled joins, using a new notion of
"extended" replay paths to fix the issues with the original remapped
upquery design that prevented it from working with nodes that generate
columns from themselves. This includes tweaking replay path
generation and replay execution to make extended replay paths work
properly, plus bringing back the column_source and handle_upquery
implementations for Join that had been removed earlier.

There's a lot more information about the actual design here in a new
section I've added to the replay paths page in docs/, plus a *ton* of
comments inline in the code - see that for a detailed explanation of how
all this works.

Fixes: ENG-1120
Change-Id: I02ea64c1810a96e487762ad538d40d8fba901f8a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2064
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Fran Noriega <fran@readyset.io>