ops/cfn: implement recommended disk configurations for consul server
Updating the Consul authority CFN template to bring our root EBS
volume's configurations up to the recommended specs provided by
HashiCorp for Consul clusters.

Consul and Zookeeper are both very sensitive to disk latency. A modest
increase of 5-10ms can cause servers to unexpectedly be flagged as
unhealthy, which of course can be remediated but it can cause erratic
performance and errors.

Here's some docs that detail the EBS configurations implemented as
default values in this CL:

https://bit.ly/36cMuFD

Change-Id: If17e42d92d6083f4572610a035824372a9c889d2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2106
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>