domain: Factor out replay paths into their own structure
As the first step in what will likely be a very long road of separating
out the various bits of the Domain structure into more isolated,
unit-testable components, factor out a data structure for storing
information about replay paths, tags, and generated columns within a
domain, providing methods to query the structure based on the ways we
were doing so throughout the domain. This incidentally gets us a little
bit of deduplication, which is nice, but more importantly lets us start
writing unit tests covering the complexity of some of this logic!
I've *started* doing that here, but as we refactor this further and add
more complexity in the future we'll likely want to write even more of
those.

Change-Id: I165be07ba4a4e68bf10a9e5502da7ef655c54743
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2107
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>