noria-mysql: Add a vertical test for pagination
Add a vertical test for a simple grouped pagination query, with randomly
generated page number. Passed first try! (which maybe should be
suspicious, but I looked at the actual generated operations here and it
looked like it was indeed exercising the functionality properly).

Getting this working required a slight refactor to the vertical testing
framework itself; since this is the first instance of a key for a query
that doesn't directly come from a column in a table, I had to add
support for that, which also involved removing the (tbh probably
misguided) static-length arrays for key columns in the operations;
overall I think this is a simplification, and it's not like we really
hurt from the extra heap allocation here since this is just a test.

Fixes: ENG-1123
Change-Id: I6dac327c02c4ac1acd73b6fd2d146c8b56a2c94e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2109
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>