query_generator: Add a new "constant" ColumnGenerationSpec
Add a new ColumnGenerationSpec::Constant variant, which has a single
DataType which is always generated for that column. This is parsed from
a string into DataType::Text in the from_str impl, then converted to the
actual target type when converting to a ColumnGenerator (we already had
a ColumnGenerator::Constant, so this just uses that).

Change-Id: Ica9121b37f511e6fc3786876c1c16968848c6e3e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2114
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>