Create Solidus pgsql demo
Creates a version of the solidus demo that connects to readyset-psql.
Creating this version of the demo mainly required swapping out the
database driver built by the solidus ruby application, and changing the
database configuration to point to the readyset-psql instance.

This commit creates a mysql and pgsql subdirectory in solidus,
which each contain the entire environment for the respective demo.

To bring this demo up:
1. Use docker-compose to bring up the postgres database.
('docker-compose up -d postgres')
2. (using the nix configuration) run 'rake db:create db:migrate db:seed'
to initialize the database.
3. Run 'rake spree_sample:load' to load the database.
4. Run 'rails s -b 0.0.0.0'. At this point you should be able to test
that the solidus demo application is running without readyset. The '-b'
option is included to allow connections from any IP address.
5. Close the Solidus application.
6. Use docker-compose to bring up the readyset-psql instance.
('docker-compose up -d db').
7. Bring up the rails application again with 'rails s -b 0.0.0.0'.

There are currently several errors preventing the demo from working
correctly. However, at this point you should be able to inspect the logs
for the server and adapter to see that replication is not working, and
the adapter is unable to process certain commands from the application.
Additionally, the application should fail with an error from readyset.

Note: config/database.yml may need updating. If solidus attempts to
connect to upstream directly instead of readyset, take a look there.

Change-Id: I7da4aa1c7b2d42a6f8057a670d9b59daf30f8a9a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2123
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>