noria_logictest: Add an --exclude option to permute
Allow generating permutations of logictests *excluding* a given set of
operators, to allow skipping generating combinations of operators that
we don't support (eg aggregates on top of IN, etc.).

Since the code to actually generate these permutations is getting a
little complex, this also extracts it out into a separate struct and
writes a couple of unit tests for this new behavior.

Change-Id: Iad5b05741efb2ed425c525348b6ef8ae3eeb585b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2122
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>