noria-client: Make set-statement-allowed a method of QueryHandler
Whether or not a SET statement should be allowed is a different question
based on the type of upstream database we're using, but previously the
logic we had for whether or not to allow SET statements was basically
entirely mysql-specific. To enable making that logic different for
postgres vs mysql, this introduces a new is_set_allowed method on the
QueryHandler trait, and migrates all the existing logic to the impl of
that method on the mysql QueryHandler. The postgres version of this
method is currently stubbed out to always return false, but soon will be
updated to have postgres-specific logic.

Change-Id: Id10299822fb8c8008256ec3e7d01919f69239e20
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2125
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>