vertical: Append extra_key_strategies to existing_keys
Turns out the reason this wasn't working was because we were appending
the extra_key_strategies (used to generate the parameter for the page
number, since it doesn't come from a column in a table) to the *new*
keys we generated, but not to the keys generated based on *existing*
rows in the database - when run at a high enough test-case count, this
caused enough failures in mysql that proptest (correctly) decided
something must be wrong with the test, but at a lower test count we'd
just only generate tests with new keys, not existing keys.

This fixes that, and also de-ignores the paginate_grouped test as it's
now passing.

Change-Id: If2a14e6fe8eb11a37321315e57750644f1e397c4
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2129
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>