nom-sql: Refactor Variable to be a struct
Rather than Variable being an enum where each member has a single
SqlIdentifier field, make it a struct with a `scope` field (of a new
VariableScope enum) and a `name` field. This both makes it easier to
deal with, as we can get at things like the name without having to
pattern-match on it, and allows us to de-duplicate some of the scope
parsing code since we now have a scope enum that we can return.

Change-Id: Ib39b1020a5f3d87c322db0e5596300dd9d0a136e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2132
Reviewed-by: Peter Farr <peter@readyset.io>
Tested-by: Buildkite CI