nom-sql: Add parser+AST for SET of postgres parameters
The semantics of the SET statement in postgresql are pretty different
than in MySQL - postgres doesn't have variables, so the SET statement is
just used to set configuration flags at runtime. It also has pretty
substantially different syntax, enough that it feels worth separating
out the AST node for SETting a postgresql parameter from the AST node
for SETting a user variable. This does that, and implements full parse
rules for SET of postgresql parameters.

Change-Id: Ic5c1d21ab1ae1bf2165ebe42c8eba4cb1fd2bc6c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2133
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>