Panic Refactor: Remove/allow panics from migration
Adds a new ReadySetError::InvalidQuery function. Also adds a couple
helper functions for generating ReadySetErrors. The primary benefit of
this CL is converting some panics that could have occurred while
processing queries to errors. It is generally easy to recover from these
errors because we have not yet made changes to the dataflow state in
these cases.

Many panics are left in the code, but these generally have to do with
areas where we would run into an inconsistent dataflow state. It is not
clear that we would be able to recover in these cases if we were to
return an error instead.

There are still several files to go in noria-server before we can
enforce the clippy lints.

Change-Id: I6a864edb1c7ff650ac2973dd509844c48e00d902
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2136
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>