migration: Handle removal of nodes in the context of a migration.
We were sending domain messages _after_ a migration was committed (you
can check the changes made to `apply_recipe` to see that).
We are now making all the changes in the context of a migration plan,
where we create the messages to be sent to the domains.

As a side note (and opinion on my part):
The changes to make that happen are a little "hacky", since we are
breaking the encapsulation between the `Recipe` and the `DataflowState`
(we are making changes to the latter, in the context of the former),
which is confusing, as it feels like that's what the migration itself
should be doing.
There's a lot of technical debt that stems from the interactions between
 the `DataflowState`, the `Recipe` and the `Migration`s; as the logic is
 pretty convoluted and hard to read and follow; and might be simplified.
In the future, we should think about the semantics of each and what we
expect from them, and refactor the code accordingly.

Refs: ENG-1179
Change-Id: Ie9ef28dbc1a58aa826212a9ac8cc10242d1c97af
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2138
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>