noria-client: Use the original query string for fallback
When we fail to execute an ad-hoc select query on noria, use
the *original* query string sent to us by the client rather than
re-formatting the parsed query into a string. We have to do this for
postgres because all of the Display implementations for the nom-sql AST
format things in MySQL-style, using backticks to escape identifiers,
using mysql-style string literals, etc.

Change-Id: I0ff26823136423ec21eb466dfbbf6df3934acdfe
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2146
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>