psql-srv: Send server_version on startup
Rails, and likely many other clients, validates that the version of the
postgresql server it's connected to is sufficient to handle the things
it's going to want to do, and bails early if the version is either not
specified by the server or too low. The way the version works in the
protocol is the serever sends a ParameterStatus backend message to the
client as part of the startup handshake, before it sends
ready_for_query_idle - the actual *value* of the ParameterStatus message
is a string which libpq should be able to parse into a version number -
it's quite ad-hoc (it uses sscanf!) but is able to handle basically any
two- or three-segment dotted version number at the start of the string,
followed by anything at all (the running version of postgres I've got in
a docker container sends the string "13.4 (Debian 13.4-1.pgdg100+1)" at
startup).

This commit sends that server version at startup, setting it to be 13.4
since that's what we're testing with, followed by "(ReadySet)" so
clients know they're connected to readyset.

Change-Id: I57d5daf08b1938585d3161966966c42295189977
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2147
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>