replicators: Properly escape column names
We were previously escaping the name of the table sent to readyset, but
not the names of the columns, either in the column list itself or the
constraint list. This meant that any table with a reserved keyword in
the name of its column (like `key`, in my case) would fail to replicate.
Doing this for the column list is trivial, since we have the name of
the column and its type as separate fields, but doing it for the
constraint list is more involved, since the only thing we get back from
postgres is the full original source of the constraint definition (the
problem is that this may succeed in parsing with the postgresql dialect
but fail to parse with the mysql dialect). The approach I took here is
to just *parse* the constraint definition using the postgresql dialect,
then re-format it using the Display impl (which uses the MySQL dialect,
which noria can parse). This works quite well, and any performance
implications are likely negligible since this only happens during
snapshotting.

Change-Id: I108fcc7bb36b5f87dcccbac44df82da18d86e132
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2149
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>