replicators: constraints can have null columns
With the way the query to load the list of constraints on a table was
written previously, it was possible for both the second and third
column (the constraint definition and the constraint kind, respectively)
to be null - in my case it looks like this was the case for a non-unique
index - but the ConstraintEntry struct wasn't set up to handle either of
those cases!

We don't want any constraints without a definition (since we don't have
anything to do with them), so we can just filter those out - and nothing
is using the constraint kind field right now, so we can just make it an
Option without any damage.

Change-Id: Icfb8707a996db4fcbd5c9e3aa15348717ea836c9
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2151
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>