ops/buildkite: extend functionality of dynamic buildkite agent logic
Adding logic to allow for customization of min_size for buildkite agent
queues, so we can prewarm a few agents vs always going to 0 on cheaper
instance types.

Also increasing the ubound of some of our queues, as we're growing as a
team very quickly! Autoscaling should ensure we're only using what we
need.

Change-Id: I724a085fb34f0ba6db733c773f70e432e4890975
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2157
Reviewed-by: Harley Klein <harley@readyset.io>
Terraform-Plan-Review: Harley Klein <harley@readyset.io>
Tested-by: Buildkite CI