replicators: Use the DataType FromSql impl directly
I think this was written before we had a FromSql implementation for
DataType, which is why it manually matches on the type of the column and
does the conversion directly rather than using the existing logic in
that impl - regardless, the FromSql impl is more complete, most notably
supporting NULL values, which the manual matching logic didn't. This
switches over to use that, which both reduces some duplication and gives
us support for replicating columns with NULL, the latter of which is
covered by some new values in the replication integration test.

Change-Id: Ic34c4acb7e1760ff1e85fc44643f6b10024e5d91
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2160
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>