ops/cfn: Add ExtraEnvironment parameters for server+adapter
Add ReadySetServerExtraEnvironment and ReadySetAdapterExtraEnvironment
parameters to both the mysql and postgresql templates, which allow
providing a comma-separated list (this prevents env vars with commas
sadly, but I don't think that's going to be an issue and cloudformation
doesn't support multiline strings being provided in the UI because it
sucks) of extra environment variables, which eventually get spliced in
to the EnvironmentFile for the relevant service.

This is something that's going to be generally useful for us as we add
new configuration via env vars to the various processes, but for now
I'm adding this to allow us to enable the feature flag for pagination in
the cloudformation-based benchmarks without having to expose it directly
to users (since it's a semi-hidden feature flag due to pagination not
yet being fully supported).

Change-Id: Icc8ad1ecbd6ad2e19f0b4a968cf61d8a9c07f153
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2164
Tested-by: Buildkite CI
Reviewed-by: Marcus Volhontseff <marcus.v@readyset.io>
Reviewed-by: Marcus Bastian <marcus@readyset.io>