Tweak all CACHED QUERY syntax
The Decision Has Been Made that we're making the following changes to
the syntax around CACHED QUERY:

- CREATE CACHED QUERY <name> AS <statement>
  becomes
  CREATE CACHE <name> FROM <statement>

- CREATE CACHED QUERY <name> AS <qid>
  becomes
  CREATE CACHE <name> FROM <qid>

- DROP CACHED QUERY <name>
  becomes
  DROP CACHE <name>

- SHOW CACHED QUERIES
  becomes
  SHOW CACHES

it has Been Decided that SHOW PROXIED QUERIES should stay unchanged.

With all of this, I've also updated the names of the various enum
variants and structs to maintain consistency with the actual syntax
they're representing.

Change-Id: Id9fa2ab4cbb6c6c5269d2f172bfada7c9c3d9a2f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2167
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>