benchmarks:  Allow queries to be specified directly on the command-line
Previously, they were required to be supplied in a file.  This can be
good for reuse, for example in CI, but for local testing it can be
inconvenient.

This also switches things around to load file-supplied queries from disk
eagerly, which saves some disk I/O and - far more importantly - allows
us to return an error to the user much sooner.

Change-Id: I3562d7db97f551e569579d1944622c52c2f5ef25
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2170
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>