readers: Don't clobber earlier reads with split ranges
with partial BETWEEN, it can end up being the case that a query for a
range ends up getting split into *multiple* upqueries, if there are
holes filled that are fully covered by that range. Previously, if that
happened, we'd accidentally overwrite the result set for the entire
original key with any results received for each upquery key. The
problematic code looked something like:

    read[read_i] = rs;

This previously made the assumption, due to having been designed for
point queries instead of range queries, that a key's position in a read
reply batch was either done, or it had a placeholder value in it, and
didn't account for keys that we want to add new rows to

Fixing this requires a pretty major rework of the way we store
partially-complete reads in the blocking read worker - previously, to
avoid having to clone any data borrowed from inside a live reader_map
handle, we'd serialize result sets as soon as we successfully read them
- but since (at least currently) our serialization mechanism is totally
opaque, this doesn't give us the ability to add new rows to those result
sets! Avoiding those clones is still probably something we want to do
though, so the approach this commit takes is to introduce a new
ServerReadReplyBatch enum, which can either contain pre-serialized bytes
in the case of no-miss points or ranges where we know we're not waiting
for more data, or cloned rows in the case where we've missed on only
part of a range, or Empty in the case where we know we're waiting on
entirely new information. This is quite messy, but probably worth it
given this is literally the most performance critical part of the
codebase.

I've run the aba-benchmark.sh script on this change, and it looks like
there's no notable regression in performance:

Baseline:

  QueryBenchmark:
    query:
      query:
        file: irl/irl_query10.sql
  Iteration 1 Results:
    query_duration (Microseconds - Decreasing goal)
    - Samples: 1540595
    - Min: 306
    - Max: 1061887
    - Mean: 2470.249393903007
  QueryBenchmark:
    query:
      query:
        file: fastly/fastly_read_query.sql
  Iteration 1 Results:
    query_duration (Microseconds - Decreasing goal)
    - Samples: 18023
    - Min: 179
    - Max: 45279
    - Mean: 1610.0962658824828

With this change:

  QueryBenchmark:
    query:
      query:
        file: irl/irl_query10.sql
  Iteration 1 Results:
    query_duration (Microseconds - Decreasing goal)
    - Samples: 1543944
    - Min: 343
    - Max: 1047551
    - Mean: 2467.283558859651
  QueryBenchmark:
    query:
      query:
        file: fastly/fastly_read_query.sql
  Iteration 1 Results:
    query_duration (Microseconds - Decreasing goal)
    - Samples: 18027
    - Min: 193
    - Max: 46335
    - Mean: 1587.2559494092227

Change-Id: Id0cb8cb9b27ec3f3000338a0f77d9f2798348aaa
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2179
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>