noria: Remove `install_recipe` method.
The `install_recipe` method was used to clean up the current state and
install the queries that where passed as a parameter.
Currently, the Noria server is meant to consume changes on the
underlying schemas of the actual database, so such functionality is
unnecessary.
All `install_recipe` appearances were replaced by `extend_recipe` (we
should do some rename work, to stop using "recipe", "soup",
"ingredients", etc; in this case, maybe rename `extend_recipe` to
`process_change` or something easier to understand).

This also fixes a parsing and handling issue we had with `DROP CACHE`
statements.

Refs: ENG-1196
Change-Id: I4edaad4290b03aea0f961dd1de6b212cad1f02a0
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2207
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>