Detect special pub/priv subnet case
There's a common case where a user has exactly one public and one
private subnet in each availability zone. We can help such customers
out by detecting such a case, and giving them the option to simply
select whether they would like to use their public, or private subnets.

This commits adds this functionality. It is done by looking up the route
tables, which are attached to a subnet. For a public subnet, there will
be a route in its route table that directs to an internet gateway. An
internet gateway is always prefixed by "igw" for the gateway id. We can
use this information to find all public subnets, and map them back
according to user intent.

Change-Id: I1d6e7d56ce3c62d77c5238008281a2293b28d976
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2180
Tested-by: Buildkite CI
Reviewed-by: Jay Bensal <jay@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>