ops/helm: introduced non-root, unified readyset-adapter container image
To simplify our Helm chart, future benchmarking automation, and other
similar scenarios, this CL aims to provide a new container image for the
adapter. Instead of needing to conditionally switch between two images,
we can choose which adapter to use at run-time.

The new Dockerfile builds both the psql and mysql adapters, and provides
a docker entrypoint script.

When this entrypoint is called, $ENGINE environment variable is checked
to determine which adapter binary to run.

Combining the two binaries into a single container image increases the
size of the image by ~100 MB (so, fairly benign).

I've also configured the new Dockerfile to use a non-root user, so we're
also moving the needle a bit on the security side of the house with this
changeset.

Change-Id: Iaa7c8268cba8ae6d4fc9ebeda5bdb784a4cbd8b0
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2182
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>