installer: re-prompt for db creds
When continuing an existing AWS deployment, we check
for existing database credentials in the deployment. This
commit adds in a prompt to either use the already configured
credentials (as we do today) or allow the user to set new ones.

If a user forgets their previous database password or wants
to change it, then they don't have to wipe out the whole
deployment.

Change-Id: Ib799324a14150bbf116b05f4438f8ba46e4845ad
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2184
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>