ops/docker: adding readyset-server container entrypoint script for k8s
This CL adds a docker-entrypoint for ReadySet-Server that can be used to
probe Consul for its' latest leader election status. This can be useful
in multiple scenarios but in particular in k8s, where we want our
components to come up in a certain order. Specifically, we don't want to
launch ReadySet-server if Consul isn't available.

In this entrypoint, CONSUL_HEALTH_SLEEP and CONSUL_HEALTH_MAX_RETRY
default to what I consider to be ideal values for k8s, whereby we "fail
fast" but not too fast (which could cause the pod to enter a
CrashLoopBackoff).

Change-Id: Ie1fcddaf165b07d3cd4692c53e947432bfd534c9
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2185
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>