noria: Standardized the input for `extend_recipe`.
We were using the raw pseudo-SQL queries (in MySQL dialect) as input for
 `extend_recipe`, which was making it hard to maintain the `Recipe`
state, since:
1. The `Recipe` (and consequently, the `extend_recipe` method) can't
handle all queries, but only a small subset; so accepting all of them
was making it harder to understand how to handle the state.
The `Recipe` should only be handling the following statements:
- CREATE TABLE
- CREATE CACHED QUERY
- CREATE VIEW
- DROP CACHED QUERY
- DROP TABLE
2. We were using pseudo-SQL queries, `VIEW x: ...` and `QUERY x: ...`,
which again made things harder to parse and understand.
Those two should have been `CREATE VIEW x AS ...` and `CREATE CACHED
QUERY x AS ...`, respectively.

This commit then does the following:
1. It moves the `changelist` module into Noria, so it can be used by the
 server and clients (to parse the queries).
2. The `Change`s struct was modified from `Add`/`Remove`, to have the
different _real_ changes we can have (what the `Recipe` can accept,
basically).
3. The old `SqlExpression` in the `recipe/mod.rs` module was changed to
`RecipeExpression`, which is meant to represent only the expressions
that the `Recipe` should *store* (we don't store `DROP TABLE` or `DROP
CACHED QUERY` statements).
4. We replaced the use of `extend_recipe(raw_query)` for `extend_recipe
(ChangeList::from_raw_queries(raw_query, dialect).unwrap())`
throughout the repo (and we specify the dialect so the Postgres adapter
can correctly parse their queries).
5. We transformed the `VIEW x: ...` and `QUERY x: ...` statements for
`CREATE VIEW x AS ...` and `CREATE CACHED QUERY x AS ...` throughout the
 repo.

There were also some improvements in the way we are parsing a set of
statements.
We are now using nom-sq for that, although we still perform the parsing
in two steps: first we separate each statement into its own string, and
then we parse them (so that if one fails, we know exactly which one was
it).

Refs: ENG-1196
Change-Id: I3873402d7664c5c49901d4f7e688b179bc429e2f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2208
Reviewed-by: Griffin Smith <griffin@readyset.io>
Tested-by: Buildkite CI