mysql-srv: Parse (nearly) full HandshakeResponse41
Parse more of the client's initial handshake response, specifically
the (optional) database name and auth_plugin_name, in preparation for
adding support for auth plugins.

Also refactors this whole module to use a style more close to how we use
nom elsewhere in the codebase (specifically `use`ing parse functions at
the top of the module rather than inline in the functions).

This parses all of the HandshakeResponse41 except for the
CLIENT_CONNECT_ATTRS, which we currently have no use for, and have no
need to support since the client only conditionally sends them based on
a capability flag.

Refs: ENG-1199
Change-Id: I56995218c9f80a1f93d60d79e4824bde9c5ce785
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2188
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>