mysql-srv: Switch auth plugin if necessary
Despite what's documented in the official mysql documentation[0], the
official mysql client does *not* actually default to "Old Password
Authentication", also known as "mysql_native_password", if the server
doesn't advertise that it supports CLIENT_PLUGIN_AUTH - instead, if the
client can't figure out an auth plugin based on what the server sends,
it defaults to caching_sha2_password[1]! The mariadb-client, of course,
implements the correct, documented behavior.

This means that in order to allow the official mysql client to connect
to us, we have to "support" pluggable authentication, at least at the
binary protocol level - this means sending CLIENT_PLUGIN_AUTH in the
capability flags of the initial server handshake, then looking at the
auth plugin sent by the client and, if it's different than
"mysql_native_password", sending the "Authentication Method Switch
Request" packet, receiving the response containing the auth data, and
continuing the authentication process from there.

With this change, it's now possible for both the official mysql client
and the mariadb client to connect to a mysql-srv server, though
in *both* cases we eat the cost of another pair of packets on each
connection, since the actual name of the default auth plugin that the
mariadb client uses is "dummy_fallback_auth".

[0]: https://dev.mysql.com/doc/internals/en/authentication-method.html
[1]: https://bit.ly/3JkXXjQ

Fixes: ENG-1199
Change-Id: Iff8a6dc5ca396b2d1a16bacef1b545b5131008a2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2191
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>